requirements.txt
flask
pynput
pytest
Flask-HTTPAuth
python-dotenv
